+++ modules-template/prov-template-compiler/src/main/java/org/openprovenance/prov/template/compiler/CompilerBeanEnactor2WithPrincipal.java
@@ -13,6 +13,7 @@ import org.openprovenance.prov.template.compiler.configuration.TemplatesProjectC
 import javax.lang.model.element.Modifier;
 
 import static org.openprovenance.prov.template.compiler.ConfigProcessor.*;
+import static org.openprovenance.prov.template.compiler.sql.CompilerSqlIntegration.BIFUN;
 
 public class CompilerBeanEnactor2WithPrincipal {
     private final CompilerUtil compilerUtil;
@@ -78,6 +79,7 @@ public class CompilerBeanEnactor2WithPrincipal {
          */
 
         builder.addField(inputProcessorClass,"checker",Modifier.FINAL, Modifier.PRIVATE);
+        builder.addField(BIFUN,"postProcessing",Modifier.FINAL, Modifier.PRIVATE);
 
         // add Field   static private final ThreadLocal<String> principal= ThreadLocal.withInitial(() -> "unknown");
         final TypeName THREAD_LOCAL_STRING=ParameterizedTypeName.get(ClassName.get(ThreadLocal.class), ClassName.get(String.class));
@@ -108,12 +110,14 @@ public class CompilerBeanEnactor2WithPrincipal {
         MethodSpec.Builder cbuilder3= MethodSpec.constructorBuilder()
                 .addModifiers(Modifier.PUBLIC)
                 .addParameter(ENACTOR_IMPLEMENTATION_TYPE, Constants.REALISER)
-                .addParameter(inputProcessorClass, "checker");
+                .addParameter(inputProcessorClass, "checker")
+                .addParameter(BIFUN, "postProcessing");
         compilerUtil.specWithComment(cbuilder3);
 
         cbuilder3
                 .addStatement("this.$N = $N", Constants.REALISER, Constants.REALISER)
-                .addStatement("this.$N = $N", "checker", "checker");
+                .addStatement("this.$N = $N", "checker", "checker")
+                .addStatement("this.$N = $N", "postProcessing", "postProcessing");
 
         builder.addMethod(cbuilder3.build());
 
@@ -137,7 +141,7 @@ public class CompilerBeanEnactor2WithPrincipal {
                 mspec.addStatement("return $N.generic_enact(new $T(),bean,\n" +
                         "                b -> checker.process(b),\n" +
                         "                (sb,b) -> new $T(sb,$N.get()).process(b),\n" +
-                        "                (rs,b) -> $N.beanCompleterFactory(rs).process(b))", Constants.REALISER, outputClassName, queryInvokerClass, principalVar, Constants.REALISER);
+                        "                (rs,b) -> $N.beanCompleterFactory(rs,postProcessing).process(b))", Constants.REALISER, outputClassName, queryInvokerClass, principalVar, Constants.REALISER);
 
             builder.addMethod(mspec.build());
         }