+++ RELEASES.md
@@ -5,6 +5,8 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-template-compiler: record index with principal
 * service-template: simple and composite queries with access control
 * service-template: queries with access control, incl browsing and navigation
+* service-template: hash of record, post-processing of record
+
 
 `2.0.7`
 * prov-jsonld: fixed context url